Backend/requirements/local.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
Pillow==10.1.0
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0